requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-elasticsearch-dsl==8.0.0
elasticsearch-dsl==8.11.0
Pillow==10.2.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7
pytest==7.4.4
pytest-django==4.7.0
django-filter==23.5
django-elasticsearch-dsl-drf==0.22.5
Faker==20.1.0
elasticsearch==8.11.0
setuptools
requests
django-jazzmin==2.6.0
